+++ openstl/datasets/__init__.py
@@ -1,5 +1,6 @@
 # Copyright (c) CAIRI AI Lab. All rights reserved
 
+from .dataloader_human import HumanDataset
 from .dataloader_kitticaltech import KittiCaltechDataset
 from .dataloader_kth import KTHDataset
 from .dataloader_moving_mnist import MovingMNIST
@@ -10,6 +11,7 @@ from .dataset_constant import dataset_parameters
 from .utils import create_loader
 
 __all__ = [
-    'KittiCaltechDataset', 'KTHDataset', 'MovingMNIST', 'TaxibjDataset', 'ClimateDataset',
+    'KittiCaltechDataset', 'HumanDataset', 'KTHDataset', 'MovingMNIST', 'TaxibjDataset',
+    'ClimateDataset',
     'load_data', 'dataset_parameters', 'create_loader',
 ]