+++ openstl/utils/parser.py
@@ -44,7 +44,7 @@ def create_parser():
 
     # dataset parameters
     parser.add_argument('--batch_size', '-b', default=16, type=int, help='Training batch size')
-    parser.add_argument('--val_batch_size', '-vb', default=4, type=int, help='Validation batch size')
+    parser.add_argument('--val_batch_size', '-vb', default=16, type=int, help='Validation batch size')
     parser.add_argument('--num_workers', default=4, type=int)
     parser.add_argument('--data_root', default='./data')
     parser.add_argument('--dataname', '-d', default='mmnist', type=str,
@@ -52,6 +52,8 @@ def create_parser():
     parser.add_argument('--pre_seq_length', default=None, type=int, help='Sequence length before prediction')
     parser.add_argument('--aft_seq_length', default=None, type=int, help='Sequence length after prediction')
     parser.add_argument('--total_length', default=None, type=int, help='Total Sequence length for prediction')
+    parser.add_argument('--use_prefetcher', action='store_true', default=False,
+                        help='Whether to use prefetcher for faster data loading')
 
     # method parameters
     parser.add_argument('--method', '-m', default='SimVP', type=str,