+++ tools/prepare_data/download_mmnist.sh
@@ -8,9 +8,11 @@ cd data/moving_mnist
 wget http://www.cs.toronto.edu/~nitish/unsupervised_video/mnist_test_seq.npy
 wget http://yann.lecun.com/exdb/mnist/train-images-idx3-ubyte.gz
 
+echo "finished"
 
-# download and arrange them in the following structure:
-# SimVPv2
+
+# Download and arrange them in the following structure:
+# OpenSTL
 # └── data
 #     ├── moving_mnist
 #     │   ├── mnist_test_seq.npy