+++ tools/prepare_data/download_weatherbench.sh
@@ -10,9 +10,16 @@ wget "https://dataserv.ub.tum.de/s/m1524895/download?path=%2F5.625deg%2F2m_tempe
 # down 1.40625deg `2m_temperature` (128x256) and place them in `data/weather/` according to `https://github.com/pangeo-data/WeatherBench`
 # wget "https://dataserv.ub.tum.de/s/m1524895/download?path=%2F1.40625deg%2F2m_temperature&files=2m_temperature_1.40625deg.zip" -O 2m_temperature_1.40625deg.zip
 
-# download and arrange them in the following structure:
-# SimVPv2
+echo "finished"
+
+
+# Download and arrange them in the following structure:
+# OpenSTL
 # └── data
 #     ├── weather
 #     │   ├── 2m_temperature
+#     │   ├── 10m_u_component_of_wind
+#     │   ├── 10m_v_component_of_wind
+#     │   ├── relative_humidity
+#     │   ├── total_cloud_cover
 #     │   ├── ...